backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
google-ai-generativelanguage==0.4.0
google-api-core==2.22.0
google-auth==2.36.0
google-generativeai==0.3.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.62.3
gTTS==2.5.1
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
pymongo==4.6.2
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.0.2
regex==2024.11.6
reportlab==4.1.0
requests==2.32.3
rsa==4.9
safetensors==0.4.5
seaborn==0.13.2
setuptools==75.4.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.1
starlette==0.36.3
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.27.1
websockets==12.0
wheel==0.45.0